backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.34.0,<0.35.0
pymongo>=4.11.0,<4.12.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=2.10.0,<2.11.0 